# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ cmis/src/main/java/ru/doccloud/cmis/server/repository/FileBridgeRepository.java
@@ -505,10 +505,12 @@ public class FileBridgeRepository {
 
         // get the file and parent
         DocumentDTO doc = getDocument(objectId.getValue());
-        DocumentDTO parent = getFirstParent(doc.getId());
-
-        crudService.deleteLink(parent.getId(), doc.getId());
-        Link link = crudService.addLink(Long.parseLong(targetFolderId), doc.getId());
+	DocumentDTO parent = getFirstParent(doc.getId());
+	
+	if (parent!=null){
+        	crudService.deleteLink(parent.getId(), doc.getId());
+	}
+	Link link = crudService.addLink(Long.parseLong(targetFolderId), doc.getId());
 
 
         return compileObjectData(context, doc, null, false, false, userReadOnly, objectInfos);
